Revert "same file callback"
This reverts commit 8944e3885150471fd64865941836fe89cb56065a.
